backend/requirements.txt
websockets
pymongo
python-dotenv
google-generativeai